codex warning clean

+++ mindspore/lite/src/common/ops/populate/reverse_populate.cc
@@ -42,7 +42,7 @@ OpParameter *PopulateReverseParameter(const void *prim) {
     free(param);
     return nullptr;
   }
-  param->num_axis_ = flatAxis->size();
+  param->num_axis_ = static_cast<int>(flatAxis->size());
   if (param->num_axis_ > REVERSE_SHAPE_MAX_SIZE) {
     MS_LOG(ERROR) << "Invalid axis size: " << param->num_axis_;
     free(param);
